Move 'new_search' to 'search' directory
