+++ lib/search/document/document.ts
