+++ lib/search/document/location.ts
