+++ lib/search/document/structure.ts
