+++ lib/search/encoding/load.ts
