+++ lib/search/encoding/store.ts
