+++ lib/search/encoding/structure/decoder.ts
