+++ lib/search/encoding/structure/encoder.ts
