+++ lib/search/index/builder.ts
