+++ lib/search/index/prepared.ts
