+++ lib/search/index/stem.ts
