+++ lib/search/index/stop.ts
