+++ lib/search/index/tokens.ts
