+++ lib/search/tree/node.ts
